+++ epub-reader.el
@@ -3,7 +3,7 @@
 ;; Copyright (C) 2026 chenyibin
 ;; SPDX-License-Identifier: GPL-3.0-or-later
 ;; Author: chenyibin
-;; Version: 0.1.0
+;; Version: 0.2.0
 ;; Package-Requires: ((emacs "29.1") (textui "0.5.1"))
 ;; Keywords: multimedia, hypermedia